Move worker to tagstore, fix n-processes calc
